services/ai-service/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
qdrant-client>=1.16.2
sentence-transformers>=3.0.0
pika>=1.3.2
python-dotenv>=1.0.1
pymongo>=4.0.0
pandas==3.0.1